line following

+++ include/motor/motor_driver.h
@@ -22,7 +22,7 @@ class Motor
 public:
     Motor(uint8_t step_pin, uint8_t dir_pin);
     void setDirection(MotorDirection direction);
-    void setSpeed(uint8_t speed);
+    void setSpeed(int32_t speed);
     void changeDirection();
     uint8_t getCurrentSpeed() const;
     MotorDirection getCurrentDirection() const;
@@ -38,7 +38,7 @@ private:
     uint pwm_channel;
 
     bool enabled = false;
-    uint8_t current_speed = 0;
+    uint32_t current_speed = 0;
     MotorDirection current_dir = STOP;
 };
